requirements.txt
cvat_sdk==2.9.2
pycocotools==2.0.6
pandas==1.5.3
gspread==5.11.1
tqdm==4.65.0
python-dotenv==1.0.0